Consolidate iOS components and split oversized views
New reusable components:
- PrimaryButton: the standard brand CTA (coral, white bold label, 54pt,
  loading/disabled/destructive states), mirroring Android's PrimaryButton.
  Adopted by all four auth submit buttons, the detail add-to-cart, and
  Offline Mode's Sync Cache / Clear Cache (destructive) buttons — replacing
  seven ad-hoc inline button styles.
- EmptyStateView gained an optional action button and is now the single
  empty-state pattern: Home migrated off HomeEmptyStateView (deleted) and
  Orders' inline 'No orders yet' text became a full empty state with a
  'Browse Cocktails' CTA that switches to the Home tab (parity with Android).

View slimming (pure code moves, identical UI):
- CocktailDetailView: sections moved to Components/DetailSections.swift,
  448 -> 128 lines.
- OfflineModeView: cards converted from computed properties to standalone
  parameterized structs in Components/OfflineModeCards.swift, 338 -> 86 lines.

Cleanup: deleted unused LoadingOverlay.swift and ShimmerEffect.swift.
Xcode project regenerated via xcodegen + pod install; verified with a clean
simulator build.
