--- iosApp/CocktailCraft/Components/HomeEmptyStateView.swift
@@ -1,45 +1,0 @@
-import SwiftUI
-
-// MARK: - Supporting Views
-
-struct HomeEmptyStateView: View {
-    let icon: String
-    let title: String
-    let subtitle: String
-    let actionTitle: String?
-    let action: (() -> Void)?
-    
-    init(icon: String, title: String, subtitle: String, actionTitle: String? = nil, action: (() -> Void)? = nil) {
-        self.icon = icon
-        self.title = title
-        self.subtitle = subtitle
-        self.actionTitle = actionTitle
-        self.action = action
-    }
-    
-    var body: some View {
-        VStack(spacing: 16) {
-            Image(systemName: icon)
-                .font(.system(size: 60))
-                .foregroundColor(.secondary)
-            
-            Text(title)
-                .font(.title2)
-                .fontWeight(.semibold)
-            
-            Text(subtitle)
-                .font(.body)
-                .foregroundColor(.secondary)
-                .multilineTextAlignment(.center)
-            
-            if let actionTitle = actionTitle, let action = action {
-                Button(action: action) {
-                    Text(actionTitle)
-                        .fontWeight(.medium)
-                }
-                .buttonStyle(.borderedProminent)
-            }
-        }
-        .padding()
-    }
-}
