--- iosApp/CocktailCraft/Components/LoadingOverlay.swift
@@ -1,40 +1,0 @@
-import SwiftUI
-
-struct LoadingOverlay: View {
-    let message: String
-    
-    var body: some View {
-        ZStack {
-            Color.black.opacity(0.4)
-                .ignoresSafeArea()
-            
-            VStack(spacing: 16) {
-                ProgressView()
-                    .progressViewStyle(CircularProgressViewStyle(tint: .white))
-                    .scaleEffect(1.5)
-                
-                Text(message)
-                    .font(.headline)
-                    .foregroundColor(.white)
-            }
-            .padding(32)
-            .background(Color.black.opacity(0.8))
-            .cornerRadius(16)
-        }
-    }
-}
-
-// View modifier for easy use
-extension View {
-    func loadingOverlay(isShowing: Bool, message: String = "Loading...") -> some View {
-        ZStack {
-            self
-            
-            if isShowing {
-                LoadingOverlay(message: message)
-                    .transition(.opacity)
-            }
-        }
-        .animation(.easeInOut(duration: 0.2), value: isShowing)
-    }
-}
