+++ iosApp/CocktailCraft/Views/CocktailDetailView.swift
@@ -11,9 +11,12 @@ import SwiftUI
 /// mirrors Android's CocktailDetailScreen for design parity: header image,
 /// main info (name/rating/price/stock/cart controls/instructions),
 /// ingredients grouped by type, attribute chips, then related cocktails.
+/// The section subviews live in Components/DetailSections.swift.
 struct CocktailDetailView: View {
     let cocktailId: String
-    @State private var viewModel = CocktailDetailViewModelSKIE()
+    // Internal (not private) so the section extension in
+    // Components/DetailSections.swift can read them.
+    @State var viewModel = CocktailDetailViewModelSKIE()
     @Environment(\.isDarkMode) var isDarkMode
 
     // Distinguishes "load not started yet" from "loaded but not found"
@@ -21,7 +24,7 @@ struct CocktailDetailView: View {
     // Serializes cart mutations: the stepper reads the async-mirrored
     // cartQuantity, so a second tap before the first round-trip lands would
     // compute from stale state (lost or duplicated increments).
-    @State private var cartActionInFlight = false
+    @State var cartActionInFlight = false
 
     var body: some View {
         VStack {
@@ -121,328 +124,5 @@ struct CocktailDetailView: View {
             viewModel.state.cocktail != nil ? viewModel.error : nil,
             clear: { viewModel.clearError() }
         )
-    }
-
-    // MARK: - Main Info
-
-    private func mainInfoCard(_ cocktail: Cocktail) -> some View {
-        VStack(alignment: .leading, spacing: AppTheme.Spacing.md) {
-            Text(cocktail.name)
-                .font(.largeTitle)
-                .fontWeight(.bold)
-                .foregroundColor(AppColors.textPrimary(isDarkMode: isDarkMode))
-
-            HStack(spacing: AppTheme.Spacing.sm) {
-                if let category = cocktail.category {
-                    Text(category)
-                        .font(.subheadline)
-                        .foregroundColor(AppColors.textSecondary(isDarkMode: isDarkMode))
-                }
-
-                Spacer()
-
-                Image(systemName: "star.fill")
-                    .font(.caption)
-                    .foregroundColor(.yellow)
-                Text(viewModel.formatRating())
-                    .font(.subheadline)
-                    .foregroundColor(AppColors.textSecondary(isDarkMode: isDarkMode))
-            }
-
-            HStack {
-                Text(viewModel.formatPrice())
-                    .font(.title2)
-                    .fontWeight(.bold)
-                    .foregroundColor(AppColors.primary(isDarkMode: isDarkMode))
-
-                Spacer()
-
-                // Stock badge colored by the ViewModel's shared stock rule
-                Text(viewModel.getStockStatus())
-                    .font(.caption)
-                    .fontWeight(.medium)
-                    .padding(.horizontal, AppTheme.Spacing.md)
-                    .padding(.vertical, AppTheme.Spacing.xs)
-                    .background(viewModel.getStockStatusColor().opacity(0.15))
-                    .foregroundColor(viewModel.getStockStatusColor())
-                    .cornerRadius(AppTheme.CornerRadius.chip)
-            }
-
-            cartControls
-
-            if let instructions = cocktail.instructions, !instructions.isEmpty {
-                VStack(alignment: .leading, spacing: AppTheme.Spacing.sm) {
-                    Text(String(localized: "Preparation"))
-                        .font(.headline)
-                        .foregroundColor(AppColors.textPrimary(isDarkMode: isDarkMode))
-
-                    Text(instructions)
-                        .font(.body)
-                        .foregroundColor(AppColors.textSecondary(isDarkMode: isDarkMode))
-                }
-                .padding(.top, AppTheme.Spacing.sm)
-            }
-        }
-        .padding()
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .cardStyle()
-        .padding(.horizontal)
-    }
-
-    @ViewBuilder
-    private var cartControls: some View {
-        if viewModel.state.isInCart {
-            // Already in cart: quantity stepper; minus at 1 removes the line
-            HStack(spacing: AppTheme.Spacing.lg) {
-                Button(action: {
-                    runCartAction {
-                        if viewModel.state.cartQuantity <= 1 {
-                            await viewModel.removeFromCart()
-                        } else {
-                            await viewModel.updateCartQuantity(Int(viewModel.state.cartQuantity) - 1)
-                        }
-                    }
-                }) {
-                    Image(systemName: viewModel.state.cartQuantity <= 1 ? "trash" : "minus.circle.fill")
-                        .font(.title2)
-                        .foregroundColor(AppColors.primary(isDarkMode: isDarkMode))
-                        .minimumHitTarget()
-                }
-                .disabled(cartActionInFlight)
-                .accessibilityLabel(viewModel.state.cartQuantity <= 1 ? "Remove from cart" : "Decrease quantity")
-                .accessibilityIdentifier("detail.decrementButton")
-
-                Text("\(viewModel.state.cartQuantity) in cart")
-                    .font(.headline)
-                    .foregroundColor(AppColors.textPrimary(isDarkMode: isDarkMode))
-                    .frame(maxWidth: .infinity)
-                    .accessibilityIdentifier("detail.cartQuantityValue")
-
-                Button(action: {
-                    runCartAction {
-                        await viewModel.updateCartQuantity(Int(viewModel.state.cartQuantity) + 1)
-                    }
-                }) {
-                    Image(systemName: "plus.circle.fill")
-                        .font(.title2)
-                        .foregroundColor(canIncrementQuantity
-                            ? AppColors.primary(isDarkMode: isDarkMode)
-                            : AppColors.gray)
-                        .minimumHitTarget()
-                }
-                .disabled(!canIncrementQuantity || cartActionInFlight)
-                .accessibilityLabel("Increase quantity")
-                .accessibilityIdentifier("detail.incrementButton")
-            }
-            .padding(.vertical, AppTheme.Spacing.sm)
-        } else {
-            Button(action: {
-                runCartAction { await viewModel.addToCart(quantity: 1) }
-            }) {
-                HStack {
-                    Image(systemName: "cart.badge.plus")
-                    Text(viewModel.canAddToCart() ? "Add to Cart" : "Out of Stock")
-                        .fontWeight(.semibold)
-                }
-                .foregroundColor(.white)
-                .frame(maxWidth: .infinity)
-                .padding()
-                .background(viewModel.canAddToCart()
-                    ? AppColors.primary(isDarkMode: isDarkMode)
-                    : AppColors.gray)
-                .cornerRadius(AppTheme.CornerRadius.button)
-            }
-            .disabled(!viewModel.canAddToCart())
-            .accessibilityLabel("Add to cart")
-            .accessibilityIdentifier("detail.addToCartButton")
-        }
-    }
-
-    private var canIncrementQuantity: Bool {
-        guard let cocktail = viewModel.state.cocktail else { return false }
-        return viewModel.state.cartQuantity < cocktail.stockCount
-    }
-
-    /// One cart mutation at a time: quantities are computed from the
-    /// async-mirrored state, so overlapping mutations would read stale values.
-    private func runCartAction(_ operation: @escaping () async -> Void) {
-        guard !cartActionInFlight else { return }
-        cartActionInFlight = true
-        Task {
-            await operation()
-            cartActionInFlight = false
-        }
-    }
-
-    // MARK: - Ingredients
-
-    private func ingredientsSection(_ cocktail: Cocktail) -> some View {
-        VStack(alignment: .leading, spacing: AppTheme.Spacing.md) {
-            Text(String(localized: "Ingredients"))
-                .font(.headline)
-                .foregroundColor(AppColors.textPrimary(isDarkMode: isDarkMode))
-
-            let types = viewModel.getAllIngredientTypes()
-            if types.isEmpty {
-                // No grouping available: flat list straight off the model
-                ForEach(Array(cocktail.ingredients.enumerated()), id: \.offset) { _, ingredient in
-                    ingredientRow(name: ingredient.name, measure: ingredient.measure)
-                }
-            } else {
-                ForEach(types, id: \.self) { type in
-                    VStack(alignment: .leading, spacing: AppTheme.Spacing.xs) {
-                        Text(type)
-                            .font(.subheadline)
-                            .fontWeight(.semibold)
-                            .foregroundColor(AppColors.primary(isDarkMode: isDarkMode))
-
-                        ForEach(viewModel.getIngredientsByType(type), id: \.self) { entry in
-                            Text("• \(entry)")
-                                .font(.body)
-                                .foregroundColor(AppColors.textSecondary(isDarkMode: isDarkMode))
-                        }
-                    }
-                    .padding(.bottom, AppTheme.Spacing.xs)
-                }
-            }
-        }
-        .padding()
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .cardStyle()
-        .padding(.horizontal)
-    }
-
-    private func ingredientRow(name: String, measure: String?) -> some View {
-        Text("• \(measure.flatMap { $0.isEmpty ? nil : "\($0) " } ?? "")\(name)")
-            .font(.body)
-            .foregroundColor(AppColors.textSecondary(isDarkMode: isDarkMode))
-    }
-
-    // MARK: - Nutrition
-
-    private func nutritionSection(_ nutrition: NutritionFacts) -> some View {
-        VStack(alignment: .leading, spacing: AppTheme.Spacing.md) {
-            Text(String(localized: "Estimated Nutrition"))
-                .font(.headline)
-                .foregroundColor(AppColors.textPrimary(isDarkMode: isDarkMode))
-
-            HStack(spacing: AppTheme.Spacing.sm) {
-                nutritionStat(value: "\(nutrition.calories)", unit: "kcal", label: "Calories")
-                nutritionStat(value: "\(nutrition.alcoholGrams)", unit: "g", label: "Alcohol")
-                nutritionStat(value: "\(nutrition.carbsGrams)", unit: "g", label: "Carbs")
-                nutritionStat(value: "\(nutrition.sugarGrams)", unit: "g", label: "Sugar")
-            }
-
-            Text("Rough keyword-based estimates, not measured data.")
-                .font(.caption2)
-                .foregroundColor(AppColors.textSecondary(isDarkMode: isDarkMode))
-        }
-        .padding()
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .cardStyle()
-        .padding(.horizontal)
-    }
-
-    private func nutritionStat(value: String, unit: String, label: String) -> some View {
-        VStack(spacing: AppTheme.Spacing.xs) {
-            Text(value)
-                .font(.title3)
-                .fontWeight(.bold)
-                .foregroundColor(AppColors.primary(isDarkMode: isDarkMode))
-            Text(unit)
-                .font(.caption2)
-                .foregroundColor(AppColors.textSecondary(isDarkMode: isDarkMode))
-            Text(label)
-                .font(.caption)
-                .foregroundColor(AppColors.textSecondary(isDarkMode: isDarkMode))
-        }
-        .frame(maxWidth: .infinity)
-        .padding(.vertical, AppTheme.Spacing.sm)
-        .background(AppColors.primary(isDarkMode: isDarkMode).opacity(0.08))
-        .cornerRadius(AppTheme.CornerRadius.sm)
-    }
-
-    // MARK: - Chips
-
-    private func chipsSection(_ cocktail: Cocktail) -> some View {
-        // Attribute chips mirroring Android's DetailChipsSection
-        let chips: [(icon: String, text: String)] = [
-            cocktail.alcoholic.map { (icon: "wineglass", text: $0) },
-            cocktail.glass.map { (icon: "cup.and.saucer", text: $0) },
-            cocktail.iba.map { (icon: "rosette", text: "IBA: \($0)") }
-        ].compactMap { $0 }
-
-        return ScrollView(.horizontal, showsIndicators: false) {
-            HStack(spacing: AppTheme.Spacing.sm) {
-                ForEach(chips, id: \.text) { chip in
-                    HStack(spacing: AppTheme.Spacing.xs) {
-                        Image(systemName: chip.icon)
-                            .font(.caption)
-                        Text(chip.text)
-                            .font(.subheadline)
-                    }
-                    .chipStyle(isSelected: false)
-                }
-            }
-            .padding(.horizontal)
-        }
-    }
-
-    // MARK: - Related Cocktails
-
-    private var relatedCocktailsSection: some View {
-        VStack(alignment: .leading, spacing: AppTheme.Spacing.md) {
-            Text(String(localized: "You Might Also Like"))
-                .font(.headline)
-                .foregroundColor(AppColors.textPrimary(isDarkMode: isDarkMode))
-                .padding(.horizontal)
-
-            ScrollView(.horizontal, showsIndicators: false) {
-                HStack(spacing: AppTheme.Spacing.md) {
-                    ForEach(viewModel.state.relatedCocktails, id: \.id) { related in
-                        // Pushes another detail screen through the enclosing
-                        // stack's navigationDestination(for: String.self)
-                        NavigationLink(value: related.id) {
-                            RelatedCocktailCard(cocktail: related)
-                        }
-                        .buttonStyle(PlainButtonStyle())
-                    }
-                }
-                .padding(.horizontal)
-            }
-        }
-    }
-}
-
-// MARK: - Related Cocktail Card
-
-private struct RelatedCocktailCard: View {
-    let cocktail: Cocktail
-    @Environment(\.isDarkMode) var isDarkMode
-
-    var body: some View {
-        VStack(alignment: .leading, spacing: AppTheme.Spacing.sm) {
-            CocktailImageView(
-                imageUrl: cocktail.imageUrl,
-                height: 100,
-                cornerRadius: AppTheme.CornerRadius.sm
-            )
-            .frame(width: 140)
-
-            Text(cocktail.name)
-                .font(.subheadline)
-                .fontWeight(.medium)
-                .foregroundColor(AppColors.textPrimary(isDarkMode: isDarkMode))
-                .lineLimit(2)
-                .multilineTextAlignment(.leading)
-
-            Text(cocktail.price.asPrice)
-                .font(.subheadline)
-                .fontWeight(.bold)
-                .foregroundColor(AppColors.primary(isDarkMode: isDarkMode))
-        }
-        .frame(width: 140, alignment: .leading)
-        .padding(AppTheme.Spacing.sm)
-        .cardStyle()
     }
 }
